+++ src/classes/s-c-controller.ts
@@ -1,6 +1,14 @@
 import Sockets from "./sockets";
 import {GameSettings} from "./foundry-interfacing/game-settings";
-import {CombatPauseRules, ModuleName, SettingNames, SocketTypes, Themes, TimeKeeperStatus} from "../constants";
+import {
+    CombatPauseRules,
+    ModuleName,
+    NoteReminderNotificationType,
+    SettingNames,
+    SocketTypes,
+    Themes,
+    TimeKeeperStatus
+} from "../constants";
 import {Logger} from "./logging";
 import {CalManager, MainApplication, NManager} from "./index";
 import ConfigurationApp from "./applications/configuration-app";
@@ -10,6 +18,7 @@ import {canUser} from "./utilities/permissions";
 import GameSockets from "./foundry-interfacing/game-sockets";
 import {RoundData} from "@league-of-foundry-developers/foundry-vtt-types/src/foundry/client/data/documents/combat";
 import MultiSelect from "./renderer/multi-select";
+import {ToSeconds} from "./utilities/date-time";
 
 /**
  * The global Simple Calendar Controller class
@@ -40,7 +49,7 @@ export default class SCController {
 
     constructor() {
         this.sockets = new Sockets();
-        this.clientSettings = {id: '', theme: Themes.dark, openOnLoad: true, openCompact: false, rememberPosition: true, appPosition: {}};
+        this.clientSettings = {id: '', theme: Themes.dark, openOnLoad: true, openCompact: false, rememberPosition: true, appPosition: {}, noteReminderNotification: NoteReminderNotificationType.whisper};
         this.globalConfiguration = {
             id: '',
             version: '',
@@ -126,6 +135,7 @@ export default class SCController {
         this.clientSettings.openCompact = GameSettings.GetBooleanSettings(SettingNames.OpenCompact);
         this.clientSettings.rememberPosition = GameSettings.GetBooleanSettings(SettingNames.RememberPosition);
         this.clientSettings.appPosition = <SimpleCalendar.AppPosition>GameSettings.GetObjectSettings(SettingNames.AppPosition);
+        this.clientSettings.noteReminderNotification = <NoteReminderNotificationType>GameSettings.GetStringSettings(SettingNames.NoteReminderNotification);
     }
 
     /**
@@ -159,6 +169,7 @@ export default class SCController {
             GameSettings.SaveBooleanSetting(SettingNames.OpenCompact, clientConfig.openCompact, false).catch(Logger.error);
             GameSettings.SaveBooleanSetting(SettingNames.RememberPosition, clientConfig.rememberPosition, false).catch(Logger.error);
             GameSettings.SaveObjectSetting(SettingNames.AppPosition, clientConfig.appPosition, false).catch(Logger.error);
+            GameSettings.SaveStringSetting(SettingNames.NoteReminderNotification, clientConfig.noteReminderNotification, false).catch(Logger.error);
             //Save the global configuration (triggers the load function)
             GameSettings.SaveObjectSetting(SettingNames.GlobalConfiguration, gc)
                 .then(() => GameSockets.emit({type: SocketTypes.mainAppUpdate, data: {}}))
@@ -175,14 +186,14 @@ export default class SCController {
      */
     public getSceneControlButtons(controls: any[]){
         if(canUser((<Game>game).user, this.globalConfiguration.permissions.viewCalendar)){
-            let tokenControls = controls.find(c => c.name === "token" );
+            let tokenControls = controls.find(c => c.name === "notes" );
             if(tokenControls && tokenControls.hasOwnProperty('tools')){
                 tokenControls.tools.push({
                     name: "calendar",
                     title: "FSC.Title",
                     icon: "fas fa-calendar simple-calendar-icon",
                     button: true,
-                    onClick: MainApplication.showApp.bind(MainApplication)
+                    onClick: MainApplication.render.bind(MainApplication)
                 });
             }
         }
@@ -212,6 +223,32 @@ export default class SCController {
             }
         }
     }
+
+    /**
+     * Checks to see if the Simple Calendar not directory should be shown and if not removes all Simple Calendar notes from the list of available journal entries.
+     * @param config
+     * @param jquery
+     * @param data
+     */
+    public renderSceneConfig(config: SceneConfig, jquery: JQuery, data: SceneConfig.Data){
+        if(!this.globalConfiguration.showNotesFolder && data.journals){
+            for(let i = 0; i < data.journals.length; i++){
+                //@ts-ignore
+                const je = (<Game>game).journal?.get(data.journals[i].id);
+                if(je){
+                    const nd = <SimpleCalendar.NoteData>je.getFlag(ModuleName, "noteData");
+                    if(nd){
+                        //@ts-ignore
+                        const option = jquery.find(`option[value='${data.journals[i].id}']`);
+                        if(option){
+                            option.remove();
+                        }
+                    }
+                }
+            }
+        }
+    }
+
 
     /**
      * Triggered when the games pause state is changed.