+++ src/index.ts
@@ -63,7 +63,7 @@ Hooks.on('ready',async () => {
         SC.initialize();
         //If we are to open the main app on foundry load, open it
         if(SC.clientSettings.openOnLoad){
-            MainApplication.showApp();
+            MainApplication.render();
         }
     }
 });
@@ -82,6 +82,7 @@ Hooks.on('renderNoteSheet', NoteSheet.SetHeight);
 Hooks.on('createJournalEntry', NManager.journalEntryUpdate.bind(NManager, 0));
 Hooks.on('updateJournalEntry', NManager.journalEntryUpdate.bind(NManager, 1));
 Hooks.on('deleteJournalEntry', NManager.journalEntryUpdate.bind(NManager, 2));
+Hooks.on('renderSceneConfig', SC.renderSceneConfig.bind(SC));
 
 
 Logger.debugMode = false;